Move getCatalystType to JdbcUtils and add new getSchema() method.
